# docs/swagger.yaml
basePath: /
definitions:
  models.Billboard:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      label:
        type: string
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Cart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CartItem:
    properties:
      base_price:
        type: number
      cart:
        $ref: '#/definitions/models.Cart'
      cart_id:
        type: string
      created_at:
        type: string
      discount_amount:
        type: number
      discount_percent:
        type: integer
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      tax_amount:
        type: number
      tax_percent:
        type: integer
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      billboard:
        $ref: '#/definitions/models.Billboard'
      billboard_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Color:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.ColorCreate:
    properties:
      name:
        type: string
      value:
        type: string
    required:
    - name
    - value
    type: object
  models.ColorResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.ColorUpdate:
    properties:
      name:
        type: string
      value:
        type: string
    required:
    - name
    - value
    type: object
  models.Image:
    properties:
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.Order:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      phone:
        type: integer
      store:
        $ref: '#/definitions/models.Store'
      store_id:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order:
        $ref: '#/definitions/models.Order'
      order_id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.OrderResponse:
    properties:
      address:
        type: string
      base_total_price:
        type: number
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      phone:
        type: integer
      store_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      color:
        $ref: '#/definitions/models.Color'
      color_id:
        type: string
      created_at:
        type: string
      discount_percent:
        type: integer
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      is_archived:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: number
      price_discount:
        type: number
      size:
        $ref: '#/definitions/models.Size'
      size_id:
        type: string
      stock:
        type: integer
      store_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductCreate:
    properties:
      category_id:
        type: string
      color_id:
        type: string
      discount_percent:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      is_archived:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      size_id:
        type: string
      stock:
        type: integer
    type: object
  models.ProductResponse:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category_id:
        type: string
      color_id:
        type: string
      created_at:
        type: string
      discount_percent:
        type: integer
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      is_archived:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      price_discount:
        type: number
      size_id:
        type: string
      stock:
        type: integer
      store_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductUpdate:
    properties:
      category_id:
        type: string
      color_id:
        type: string
      discount_percent:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      is_archived:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      price:
        type: number
      size_id:
        type: string
      stock:
        type: integer
    type: object
  models.Size:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
...
  models.Store:
    properties:
      billboards:
        items:
          $ref: '#/definitions/models.Billboard'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      colors:
        items:
          $ref: '#/definitions/models.Color'
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      sizes:
        items:
          $ref: '#/definitions/models.Size'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: integer
      stores:
        items:
          $ref: '#/definitions/models.Store'
        type: array
      updated_at:
        type: string
    type: object
...
  web.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: hi.boedi8@gmail.com
    name: hi.boedi8@gmail.com
    url: https://www.github.com/hiboedi
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/{storeId}/colors:
    get:
      consumes:
      - application/json
      description: FindAll Color from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ColorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindAll Color from the store
      tags:
      - Color
    post:
      consumes:
      - application/json
      description: create Color for the store
      parameters:
      - description: Color create
        in: body
        name: Color
        required: true
        schema:
          $ref: '#/definitions/models.ColorCreate'
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ColorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: create Color for the store
      tags:
      - Color
  /api/{storeId}/colors/{colorId}:
    delete:
      consumes:
      - application/json
      description: Delete Color from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Color ID
        in: path
        name: colorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Delete Color from the store
      tags:
      - Color
    get:
      consumes:
      - application/json
      description: FindById Color from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Color ID
        in: path
        name: colorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ColorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindById Color from the store
      tags:
      - Color
    put:
      consumes:
      - application/json
      description: Update Color from the store
      parameters:
      - description: Color Update
        in: body
        name: Color
        required: true
        schema:
          $ref: '#/definitions/models.ColorUpdate'
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Color ID
        in: path
        name: colorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ColorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Update Color from the store
      tags:
      - Color
  /api/{storeId}/orders:
    get:
      consumes:
      - application/json
      description: FindAll Order from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindAll Order from the store
      tags:
      - Order
  /api/{storeId}/products:
    get:
      consumes:
      - application/json
      description: FindAll Products from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindAll Products from the store
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: create Product for the store
      parameters:
      - description: Product create
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.ProductCreate'
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: create Product for the store
      tags:
      - Product
  /api/{storeId}/products/{productId}:
    delete:
      consumes:
      - application/json
      description: Delete Product from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Delete Product from the store
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: FindById Product from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindById Product from the store
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update Product from the store
      parameters:
      - description: Product Update
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdate'
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Update Product from the store
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"